[XPU] support print runtime error log for xblas
