+++ paddle/phi/backends/CMakeLists.txt
@@ -19,7 +19,8 @@ if(WITH_GPU OR WITH_ROCM)
 endif()
 
 if(WITH_XPU)
-  list(APPEND BACKENDS_SRCS xpu/xpu_context.cc xpu/xpu_info.cc)
+  list(APPEND BACKENDS_SRCS xpu/xpu_context.cc xpu/xpu_info.cc
+       xpu/enforce_xpu.cc)
   list(
     APPEND
     BACKENDS_SRCS
