+++ paddle/phi/backends/xpu/enforce_xpu.h
@@ -27,148 +27,14 @@ namespace phi {
 namespace backends {
 namespace xpu {
 
-// Note: XPU runtime api return int, not XPUError_t
-inline const char* xpuGetErrorString(int stat) {
-  switch (stat) {
-    case XPU_SUCCESS:
-      return "Success";
-    case XPUERR_INVALID_DEVICE:
-      return "Invalid XPU device";
-    case XPUERR_UNINIT:
-      return "XPU runtime not properly inited";
-    case XPUERR_NOMEM:
-      return "Device memory not enough";
-    case XPUERR_NOCPUMEM:
-      return "CPU memory not enough";
-    case XPUERR_INVALID_PARAM:
-      return "Invalid parameter";
-    case XPUERR_NOXPUFUNC:
-      return "Cannot get XPU Func";
-    case XPUERR_LDSO:
-      return "Error loading dynamic library";
-    case XPUERR_LDSYM:
-      return "Error loading func from dynamic library";
-    case XPUERR_SIMULATOR:
-      return "Error from XPU Simulator";
-    case XPUERR_NOSUPPORT:
-      return "Operation not supported";
-    case XPUERR_ABNORMAL:
-      return "Device abnormal due to previous error";
-    case XPUERR_KEXCEPTION:
-      return "Exception in kernel execution";
-    case XPUERR_TIMEOUT:
-      return "Kernel execution timed out";
-    case XPUERR_BUSY:
-      return "Resource busy";
-    case XPUERR_USEAFCLOSE:
-      return "Use a stream after closed";
-    case XPUERR_UCECC:
-      return "Uncorrectable ECC";
-    case XPUERR_OVERHEAT:
-      return "Overheat";
-    case XPUERR_UNEXPECT:
-      return "Execution error, reach unexpected control flow";
-    case XPUERR_DEVRESET:
-      return "Device is being reset, try again later";
-    case XPUERR_HWEXCEPTION:
-      return "Hardware module exception";
-    case XPUERR_HBM_INIT:
-      return "Error init HBM";
-    case XPUERR_DEVINIT:
-      return "Error init device";
-    case XPUERR_PEERRESET:
-      return "Device is being reset, try again later";
-    case XPUERR_MAXDEV:
-      return "Device count exceed limit";
-    case XPUERR_NOIOC:
-      return "Unknown IOCTL command";
-    case XPUERR_DMATIMEOUT:
-      return "DMA timed out, a reboot maybe needed";
-    case XPUERR_DMAABORT:
-      return "DMA aborted due to error, possibly wrong address or hardware "
-             "state";
-    case XPUERR_MCUUNINIT:
-      return "Firmware not initialized";
-    case XPUERR_OLDFW:
-      return "Firmware version too old (<15), please update.";
-    case XPUERR_PCIE:
-      return "Error in PCIE";
-    case XPUERR_FAULT:
-      return "Error copy between kernel and user space";
-    case XPUERR_INTERRUPTED:
-      return "Execution interrupted by user";
-    default:
-      return "Unknown error";
-  }
-}
-
+std::string get_xpu_error_msg(int error_code);
 #ifdef PADDLE_WITH_XPU_BKCL
-inline const char* bkclGetErrorString(BKCLResult_t stat) {
-  switch (stat) {
-    case BKCL_SUCCESS:
-      return "BKCL_SUCCESS";
-    case BKCL_INVALID_ARGUMENT:
-      return "BKCL_INVALID_ARGUMENT";
-    case BKCL_RUNTIME_ERROR:
-      return "BKCL_RUNTIME_ERROR";
-    case BKCL_SYSTEM_ERROR:
-      return "BKCL_SYSTEM_ERROR";
-    case BKCL_INTERNAL_ERROR:
-      return "BKCL_INTERNAL_ERROR";
-    default:
-      return "Unknown BKCL status";
-  }
-}
+std::string get_xpu_error_msg(BKCLResult_t stat);
 #endif
-
-inline const char* xdnnGetErrorString(int stat) {
-  // Also reused by xfa and xpudnn apis.
-  switch (stat) {
-    case baidu::xpu::api::Error_t::SUCCESS:
-      return "XDNN_SUCCESS";
-    case baidu::xpu::api::Error_t::INVALID_PARAM:
-      return "XDNN_INVALID_PARAM";
-    case baidu::xpu::api::Error_t::RUNTIME_ERROR:
-      return "XDNN_RUNTIME_ERROR";
-    case baidu::xpu::api::Error_t::NO_ENOUGH_WORKSPACE:
-      return "XDNN_NO_ENOUGH_WORKSPACE";
-    case baidu::xpu::api::Error_t::NOT_IMPLEMENT:
-      return "XDNN_NOT_IMPLEMENT";
-    default:
-      return "Unknown XDNN status";
-  }
-}
-
-inline std::string build_xpu_error_msg(int stat) {
-  std::string error_msg = "XPU Error <" + std::to_string(stat) + ">, " +
-                          xpuGetErrorString(stat) + " ";
-  return error_msg;
-}
-
-#ifdef PADDLE_WITH_XPU_BKCL
-inline std::string build_xpu_error_msg(BKCLResult_t stat) {
-  std::string error_msg = "BKCL Error <" + std::to_string(stat) + ">, " +
-                          bkclGetErrorString(stat) + " ";
-  return error_msg;
-}
-#endif
-
-inline std::string build_xdnn_error_msg(int stat, std::string msg) {
-  std::string error_msg = msg + "XDNN Error <" + std::to_string(stat) + ">, " +
-                          xdnnGetErrorString(stat) + " ";
-  return error_msg;
-}
-
-inline std::string build_runtime_error_msg() {
-  auto rt_error_code = cudaGetLastError();
-  std::string error_msg = "XPU Runtime Error <" +
-                          std::to_string(rt_error_code) + ">, " +
-                          std::string(cudaGetErrorString(rt_error_code)) + " ";
-  return error_msg;
-}
+std::string get_xdnn_error_msg(int error_code, std::string msg);
+std::string get_xblas_error_msg(int error_code, std::string msg);
 
 namespace details {
-
 template <typename T>
 struct ExternalApiType {};
 
@@ -190,51 +56,57 @@ DEFINE_EXTERNAL_API_TYPE(BKCLResult_t, BKCL_SUCCESS);
 #ifdef PADDLE_WITH_XPU
 DEFINE_EXTERNAL_API_TYPE(cudaError_t, cudaSuccess);
 #endif
-
 #undef DEFINE_EXTERNAL_API_TYPE
-
 }  // namespace details
 
-#define PADDLE_ENFORCE_XPU_SUCCESS(COND)                        \
-  do {                                                          \
-    auto __cond__ = (COND);                                     \
-    using __XPU_STATUS_TYPE__ = decltype(__cond__);             \
-    constexpr auto __success_type__ =                           \
-        ::phi::backends::xpu::details::ExternalApiType<         \
-            __XPU_STATUS_TYPE__>::kSuccess;                     \
-    if (UNLIKELY(__cond__ != __success_type__)) {               \
-      auto __summary__ = common::errors::External(              \
-          ::phi::backends::xpu::build_xpu_error_msg(__cond__)); \
-      __THROW_ERROR_INTERNAL__(__summary__);                    \
-    }                                                           \
+// return code type int for xpu api, type BKCLResult_t for bkcl api
+#define PADDLE_ENFORCE_XPU_SUCCESS(COND)                      \
+  do {                                                        \
+    auto __cond__ = (COND);                                   \
+    using __XPU_STATUS_TYPE__ = decltype(__cond__);           \
+    constexpr auto __success_type__ =                         \
+        ::phi::backends::xpu::details::ExternalApiType<       \
+            __XPU_STATUS_TYPE__>::kSuccess;                   \
+    if (UNLIKELY(__cond__ != __success_type__)) {             \
+      std::string error_msg =                                 \
+          ::phi::backends::xpu::get_xpu_error_msg(__cond__);  \
+      auto __summary__ = common::errors::External(error_msg); \
+      __THROW_ERROR_INTERNAL__(__summary__);                  \
+    }                                                         \
   } while (0)
 
-#define PADDLE_ENFORCE_XDNN_SUCCESS(COND, MSG)                                 \
-  do {                                                                         \
-    auto __cond__ = (COND);                                                    \
-    if (UNLIKELY(__cond__ != baidu::xpu::api::Error_t::SUCCESS)) {             \
-      if (__cond__ == baidu::xpu::api::Error_t::RUNTIME_ERROR) {               \
-        auto __summary__ = common::errors::External(                           \
-            ::phi::backends::xpu::build_xdnn_error_msg(__cond__, MSG) + "\n" + \
-            ::phi::backends::xpu::build_runtime_error_msg());                  \
-        __THROW_ERROR_INTERNAL__(__summary__);                                 \
-      } else {                                                                 \
-        auto __summary__ = common::errors::External(                           \
-            ::phi::backends::xpu::build_xdnn_error_msg(__cond__, MSG));        \
-        __THROW_ERROR_INTERNAL__(__summary__);                                 \
-      }                                                                        \
-    }                                                                          \
+#define PADDLE_ENFORCE_XDNN_SUCCESS(COND, MSG)                     \
+  do {                                                             \
+    auto __cond__ = (COND);                                        \
+    if (UNLIKELY(__cond__ != baidu::xpu::api::Error_t::SUCCESS)) { \
+      std::string error_msg =                                      \
+          ::phi::backends::xpu::get_xdnn_error_msg(__cond__, MSG); \
+      auto __summary__ = common::errors::External(error_msg);      \
+      __THROW_ERROR_INTERNAL__(__summary__);                       \
+    }                                                              \
   } while (0)
 
-#define PADDLE_ENFORCE_XDNN_NOT_NULL(ptr)                                      \
-  do {                                                                         \
-    if (UNLIKELY(ptr == nullptr)) {                                            \
-      auto __summary__ =                                                       \
-          common::errors::External(::phi::backends::xpu::build_xdnn_error_msg( \
-              baidu::xpu::api::Error_t::NO_ENOUGH_WORKSPACE,                   \
-              "XPU memory is not enough"));                                    \
-      __THROW_ERROR_INTERNAL__(__summary__);                                   \
-    }                                                                          \
+// 由于xblas的错误类型定义源文件(xblas_api.h)不能在.h中直接引用
+// 因此使用error_msg是否为空来判断是否发生错误，返回值判断逻辑在.cc中实现
+#define PADDLE_ENFORCE_XBLAS_SUCCESS(COND, MSG)                   \
+  do {                                                            \
+    auto __cond__ = (COND);                                       \
+    std::string error_msg =                                       \
+        ::phi::backends::xpu::get_xblas_error_msg(__cond__, MSG); \
+    if (error_msg != "") {                                        \
+      auto __summary__ = common::errors::External(error_msg);     \
+      __THROW_ERROR_INTERNAL__(__summary__);                      \
+    }                                                             \
+  } while (0)
+
+#define PADDLE_ENFORCE_XDNN_NOT_NULL(ptr)                               \
+  do {                                                                  \
+    if (UNLIKELY(ptr == nullptr)) {                                     \
+      std::string error_msg = ::phi::backends::xpu::get_xdnn_error_msg( \
+          baidu::xpu::api::Error_t::NO_ENOUGH_WORKSPACE, "XPU Alloc");  \
+      auto __summary__ = common::errors::External(error_msg);           \
+      __THROW_ERROR_INTERNAL__(__summary__);                            \
+    }                                                                   \
   } while (0)
 
 #define PADDLE_ENFORCE_XRE_SUCCESS(COND)                            \
